+++ packages/elements/analytics-listeners/src/fabric/types.ts
@@ -1,0 +1,6 @@
+import { GasPayload } from '@atlaskit/analytics-gas-types';
+export type event = {
+  payload: GasPayload;
+  context: Array<{}>;
+};
+export type ListenerFunction = (event: event) => void;