+++ packages/core/global-navigation/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @atlaskit/global-navigation
 
+## 4.2.0
+
+- [minor] Add screen analytics events for navigation drawers [f02fb34](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f02fb34)
+
+
+- [none] Updated dependencies [ac34b7a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ac34b7a)
+  - @atlaskit/navigation-next@3.6.1
+- [none] Updated dependencies [f02fb34](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f02fb34)
+  - @atlaskit/navigation-next@3.6.1
+- [none] Updated dependencies [20b8844](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/20b8844)
+  - @atlaskit/analytics-listeners@3.4.0
+  - @atlaskit/navigation-next@3.6.1
+- [none] Updated dependencies [85ddb9e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/85ddb9e)
+  - @atlaskit/navigation-next@3.6.1
+  - @atlaskit/analytics-listeners@3.4.0
+
+## 4.1.6
+- [patch] Updated dependencies [b12f7e6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b12f7e6)
+  - @atlaskit/badge@9.1.1
+  - @atlaskit/navigation-next@3.5.2
+
+## 4.1.5
+- [patch] Updated dependencies [dfa100e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/dfa100e)
+  - @atlaskit/analytics-listeners@3.3.1
+  - @atlaskit/analytics-namespaced-context@2.1.1
+  - @atlaskit/navigation-next@3.5.1
+
 ## 4.1.4
 - [patch] Updated dependencies [df22ad8](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/df22ad8)
   - @atlaskit/theme@6.0.0